WIP refactoring, passing tests
