+++ parser.c
@@ -1,5 +1,5 @@
 #include <string.h>
-#include "parser.h"
+#include "Parser.h"
 
 
 // Function:    get_command_type
